CORE-680: move global file to a folder
- Add govukRebrand global nunjucks flag, but turned off
