+++ src/config/nunjucks/globals/globals.js
@@ -1,0 +1,4 @@
+// Turn on to enable GOV.UK rebrand - https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#use-the-refreshed-govuk-brand-if-youre-using-our-nunjucks-page-template
+const govukRebrand = false
+
+export { govukRebrand }
