+++ src/config/nunjucks/nunjucks.js
@@ -6,7 +6,7 @@ import hapiVision from '@hapi/vision'
 import { config } from '~/src/config/config.js'
 import { context } from './context/context.js'
 import * as filters from './filters/filters.js'
-import * as globals from './globals.js'
+import * as globals from './globals/globals.js'
 
 const dirname = path.dirname(fileURLToPath(import.meta.url))
 const nunjucksEnvironment = nunjucks.configure(
